test_wide_ep modifyed
